Tidy some worker code
----Release Notes----
[]
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=111082740

+++ sdk/src/main/java/com/google/cloud/dataflow/sdk/runners/worker/InMemoryReader.java
@@ -100,7 +100,7 @@ public class InMemoryReader<T> extends Reader<T> {
 
     @Override
     protected boolean hasNextImpl() {
-      return tracker.tryReturnRecordAt(true, (long) nextIndex);
+      return tracker.tryReturnRecordAt(true, nextIndex);
     }
 
     @Override
